requirements.txt
flask
sentence-transformers
langchain_nvidia_ai_endpoints
faiss-cpu
